backend/pyproject.toml
[tool.poetry]
name = "air_quality"
version = "0.1.0"
description = ""
authors = ["machbry <matthieu.chambery@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
sqlalchemy-bigquery = "^1.12.0"
alembic = "^1.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
